fix: config setting
