+++ src/components/modals/IncomeModal/interface.ts
@@ -1,0 +1,12 @@
+export interface IncomeModalProps {
+    title: string;
+    handleClose: () => void;
+    portalId?: string;
+    reportedIncomeSources: number;
+    reportedFinancialObligations: number;
+    subsistenceReserve: number;
+    availableForNewCommitments: number;
+    maxVacationTerm: number;
+    maxAmount: number;
+  }
+
