+++ src/components/modals/IncomeModal/stories/props.ts
@@ -1,5 +1,5 @@
 import { ArgTypes } from "@storybook/react";
-import {IncomeModalProps} from ".."
+import { IncomeModalProps } from '../interface';
 
 export const props: Partial<ArgTypes<IncomeModalProps>> = {
   title: {
